toggleTaskCompletion method moved from User to Todo
