+++ src/main/java/com/example/todolist/service/UserService.java
@@ -11,7 +11,6 @@ public interface UserService {
     User getUserById(Long userId);
     User addUser( User user);
     void addTask(Long userId, Task task);
-    void toggleTaskCompletion(Long taskId);
     User update(User newUser, Long id);
     void deleteTask(Long userId, Long taskId);
     void deleteUser(Long userId);
